Extract skill dispatcher into GameLogic::Combat
Move the skill-execution core (ExecuteSkill, CanExecuteSkill, CheckMana,
ExecuteSkillComplete, ~360 lines) out of ZzzInterface.cpp into
GameLogic/Combat/SkillExecution.{h,cpp} under namespace GameLogic::Combat,
joining the per-class executors it dispatches to. Attack() and the two
MuHelper call sites now call it qualified. No behaviour change.
