+++ src/source/Engine/Object/ZzzInterface.h
@@ -124,10 +124,6 @@ bool CheckTile(CHARACTER* c, OBJECT* o, float Range);
 void LetHeroStop(CHARACTER* c = NULL, BOOL bSetMovementFalse = FALSE);
 void SendCharacterMove(unsigned short Key, float Angle, unsigned char PathNum, unsigned char* PathX, unsigned char* PathY, unsigned char TargetX, unsigned char TargetY);
 void Attack(CHARACTER* c);
-int ExecuteSkill(CHARACTER* c, ActionSkillType Skill, float Distance);
-bool ExecuteSkillComplete(CHARACTER* c);
-bool CanExecuteSkill(CHARACTER* c, ActionSkillType Skill, float Distance);
-bool CheckMana(CHARACTER* c, int Skill);
 bool SkillElf(CHARACTER* c, ITEM* p);
 void SendRequestMagic(int Type, int Key);
 void SendRequestMagicContinue(int Type, int x, int y, int Angle, BYTE Dest, BYTE Tpos, WORD TKey, BYTE* pSkillSerial);
